feat : 보스가 피격된 무기의 공격력만큼 체력이 깍이도록 변경, 즉사기가 조금 더 느리게 발동될 수 있도록 애니메이션 조정

+++ Assets/Scripts/Entity/Boss_OMG/BossController.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 
 public class BossController : MonoBehaviour
 {
-    [Range(0, 2000)][SerializeField] private int hp = 2000;
-    public int Hp
+    [Range(0, 2000)][SerializeField] private float hp = 2000;
+    public float Hp
     {
         get => hp;
         set => Mathf.Clamp(value, 0, 2000);
@@ -123,7 +123,7 @@ public class BossController : MonoBehaviour
         }
     }
 
-    public void Damaged(int damage)
+    public void Damaged(float damage)
     {
         hp -= damage;
